pyproject.toml
[tool.poetry]
name = "home_automation_server"
version = "1.0.0"
description = "Home Automation Server"
authors = ["Kjeld Schmidt"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.1.0"
flask-apscheduler = "^1.13.1"
paho-mqtt = "^2.1.0"
apscheduler = "^3.11.0"
requests = "^2.32.3"
pytz = "^2025.1"
markupsafe = "^3.0.2"
soco = "^0.30.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-mock = "^3.14.0"
black = "^25.1.0"
mypy = "^1.15.0"
responses = "^0.25.6"
types-pytz = "^2025.1.0.20250204"
types-requests = "^2.32.0.20250301"
typing-extensions = "^4.12.2"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
mypy_path = "stubs/"
check_untyped_defs = true
disallow_untyped_defs = true
disallow_any_explicit = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_return_any = true
warn_unreachable = true
strict_equality = true
strict = true
show_error_context = true
warn_redundant_casts = true
warn_unused_ignores = true
pretty = true
